2.0

+++ MatrizWithThreads/src/main/java/Prueba/SumColumnaCallable.java
@@ -2,7 +2,7 @@ package Prueba;
 
 import java.util.concurrent.Callable;
 
-public class SumColumnaCallable implements Callable<Integer> {
+public class SumColumnaCallable implements Callable<Long> {
     private int[][] matriz;
     private int columna;
 
@@ -12,8 +12,8 @@ public class SumColumnaCallable implements Callable<Integer> {
     }
 
     @Override
-    public Integer call() {
-        int suma = 0;
+    public Long call() {
+        long suma = 0;
         for (int i = 0; i < matriz.length; i++) {
             suma += matriz[i][columna];
         }
